+++ src/utils/constant.js
@@ -1,0 +1,28 @@
+const TYPES = ['Taxi',
+  'Bus',
+  'Train',
+  'Ship',
+  'Transport',
+  'Drive',
+  'Flight',
+  'Check-in',
+  'Sightseeing',
+  'Restaurant',
+];
+const CITIES = [
+  'London',
+  'Paris',
+  'Moscow',
+  'Amsterdam',
+  'Warsaw',
+  'Minsk',
+  'Brussels',
+  'Brasilia',
+  'Ottawa',
+  'Helsinki',
+  'Berlin',
+  'Kiev',
+  'Astana',
+];
+
+export {TYPES, CITIES};
